Add a flag to hide partitions on disklabels not supported by parted.
This is mostly for anaconda. Now when there is a disklabel present
that the kernel found partitions for but parted doesn't support we
will include the partitions and whatever is on them in the
devicetree. In anaconda, the only thing visible will be the disk.
If the user removes the disklabel via recursive_remove or
initialize_disk blivet will schedule actions to remove all of the
lvm, &c on the disk before scheduling the action to remove the
disklabel.